API References

NeatText.clean_text — Method

clean_text: Clean A Given Text using a set of patterns

usepatterns: An array or collection of predefined pattern symbols - [:emails,:urls,:mentions:hashtags,:userhandles,:htmltags:puncts,:numbers,:phonenumbers,:special_char]

[:emails :urls :mentions :hashtags :userhandles :htmltags :puncts :numbers :phonenumbers :special_char]

note:

- the order of the patterns matters in how the text is cleaned

NeatText.extract_btc — Method

extract_btc: Extract all Bitcoin BTC Address/Values from a given text

NeatText.extract_currencies — Method

extract_currencies: Extract all currencies from a given text

NeatText.extract_dates — Method

extract_userhandles: Extract all dates from a given text

NeatText.extract_emails — Method

extract_emails: Extract all emails from a given text

NeatText.extract_emojis — Method

extract_emojis: Extract all emojis from a given text

NeatText.extract_hashtags — Method

extract_hashtags: Extract all hashtags from a given text

NeatText.extract_mastercards — Method

extract_mastercards: Extract all MasterCard Address/Values from a given text

NeatText.extract_md5sha — Method

extract_md5sha: Extract all MD5 or SHA Hashes from a given text

NeatText.extract_numbers — Method

extract_numbers: Extract all numbers from a given text

NeatText.extract_patterns — Method

extract_patterns: Extract Patterns from Text

extract_patterns(s,:puncts)

NeatText.extract_phonenumbers — Method

extract_phonenumbers: Extract all phone numbers from a given text

NeatText.extract_pobox — Method

extract_pobox: Extract all Post Office Box from a given text

NeatText.extract_puncts — Method

extract_puncts: Extract all punctuations from a given text

NeatText.extract_special_characters — Method

extractspecialcharacters: Extract all special characters from a given text

NeatText.extract_stopwords — Function

extract_stopwords: returns a list of stopwords in the given text param: ––-

lang: specify the language for the stopwords : en,es,fr,ru,yo,de

NeatText.extract_streetaddress — Method

extract_streetaddress: Extract all Street Address from a given text

NeatText.extract_urls — Method

extract_urls: Extract all urls from a given text

NeatText.extract_userhandles — Method

extract_userhandles: Extract all userhandles and mentions from a given text

NeatText.extract_visacards — Method

extract_visacards: Extract all VisaCards Address/Values from a given text

NeatText.fix_contractions — Method

fix_contractions: fix contractions for a given text

NeatText.remove_btc — Method

remove_btc: Remove BIT Address from a given text params: -text: text or string of text to be cleaned

NeatText.remove_creditcards — Method

remove_creditcards: Remove CreditCard numbers from a given text params: -text: text or string of text to be cleaned

NeatText.remove_currencies — Method

remove_currencies: Remove currencies from a given text params: -text: text or string of text to be cleaned

NeatText.remove_currency_symbols — Method

removecurrencysymbols: Remove currency symbols from a given text params: -text: text or string of text to be cleaned

NeatText.remove_dates — Method

remove_dates: Remove Dates from a given text params: -text: text or string of text to be cleaned

NeatText.remove_emails — Method

remove_emails: Remove emails from a given text params: -text: text or string of text to be cleaned

NeatText.remove_emojis — Method

remove_emojis: Remove emojis from a given text params: -text: text or string of text to be cleaned

NeatText.remove_hashtags — Method

remove_hashtags: Remove hashtags symbols from a given text params: -text: text or string of text to be cleaned

NeatText.remove_htmltags — Method

remove_htmltags: Remove htmltags from a given text params: -text: text or string of text to be cleaned

NeatText.remove_mastercards — Method

remove_mastercard: Remove Master Card from a given text params: -text: text or string of text to be cleaned

NeatText.remove_md5sha — Method

remove_md5sha: Remove MD5 and SHA values from a given text params: -text: text or string of text to be cleaned

NeatText.remove_numbers — Method

remove_numbers: Remove numbers from a given text params: -text: text or string of text to be cleaned

NeatText.remove_patterns — Method

remove_patterns: remove from given text your specified patterns params: -text: text or string of text to be cleaned

NeatText.remove_phonenumbers — Method

remove_phonenumbers: Remove phonenumbers from a given text params: -text: text or string of text to be cleaned

NeatText.remove_pobox — Method

remove_pobox: Remove Post Office Box Address from a given text params: -text: text or string of text to be cleaned

NeatText.remove_puncts — Method

remove_puncts: Remove punctuations from a given text params: -text: text or string of text to be cleaned

NeatText.remove_special_characters — Method

removespecialcharacters: Remove Special Characters from a given text params: -text: text or string of text to be cleaned

NeatText.remove_stopwords — Function

remove_stopwords: returns a text with all stopwords in the given text removed param: ––-

lang: specify the language for the stopwords : en,es,fr,ru,yo,de

params: -text: text or string of text to be cleaned

NeatText.remove_streetaddress — Method

remove_streetaddress: Remove street addresses from a given text params: -text: text or string of text to be cleaned

NeatText.remove_urls — Method

remove_urls: Remove URLS from a given text params: -text: text or string of text to be cleaned

NeatText.remove_userhandles — Method

remove_userhandles: Remove userhandles or mentions from a given text params: -text: text or string of text to be cleaned

NeatText.remove_visacards — Method

remove_visacard: Remove Visa Card numbers from a given text params: -text: text or string of text to be cleaned